refactor: Split functions out
