+++ src/parser.test.ts
@@ -1,109 +1,5 @@
 import * as parser from './parser.js'
 
-describe("parseAttack", () => {
-  it("handles a melee weapon attack", () => {
-    expect(parser.parseAttack("1 sword +5 (1d6)")).toEqual({
-      quantity: "1",
-      name: "sword",
-      range: undefined,
-      bonus: "+5",
-      damage: "1d6",
-    });
-  });
-  it("handles a ranged weapon attack", () => {
-    expect(parser.parseAttack("1 bow (far) +5 (1d6)")).toEqual({
-      quantity: "1",
-      name: "bow",
-      range: "far",
-      bonus: "+5",
-      damage: "1d6",
-    });
-  });
-  it("handles a thrown weapon attack", () => {
-    expect(parser.parseAttack("1 dagger (close/near) +5 (1d6)")).toEqual({
-      quantity: "1",
-      name: "dagger",
-      range: "close/near",
-      bonus: "+5",
-      damage: "1d6",
-    });
-  });
-  it("handles a spell or named trait with no bonus", () => {
-    expect(parser.parseAttack("1 spell")).toEqual({
-      quantity: "1",
-      name: "spell",
-      bonus: undefined,
-      damage: undefined,
-      range: undefined,
-    });
-    expect(parser.parseAttack("1 multi word spell")).toEqual({
-      quantity: "1",
-      name: "multi word spell",
-      bonus: undefined,
-      damage: undefined,
-      range: undefined,
-    });
-  });
-  it("handles a trait with no quantity or bonus", () => {
-    expect(parser.parseAttack("bone")).toEqual({
-      quantity: undefined,
-      name: "bone",
-      bonus: undefined,
-      damage: undefined,
-      range: undefined,
-    });
-
-    expect(parser.parseAttack("blood whip")).toEqual({
-      quantity: undefined,
-      name: "blood whip",
-      bonus: undefined,
-      damage: undefined,
-      range: undefined,
-    });
-  });
-  it("separates the bonus for spells with a bonus", () => {
-    expect(parser.parseAttack("1 spell +3")).toEqual({
-      quantity: "1",
-      name: "spell",
-      bonus: "+3",
-      damage: undefined,
-      range: undefined,
-    });
-  });
-});
-
-describe("parseAttacks", () => {
-  it("groups 'and' attacks together", () => {
-    expect(parser.parseAttacks("1 spell and 1 sword +2 (1d6)").length).toBe(1);
-  });
-  it("splits 'or' attacks up", () => {
-    expect(
-      parser.parseAttacks(
-        "1 spell or 1 sword +2 (1d6) or 1 dagger (close/near) +1 (1d4)",
-      ).length,
-    ).toBe(3);
-  });
-});
-
-describe("isTraitStart", () => {
-  it("rejects lines with no period", () => {
-    expect(parser.isTraitStart("this line lacks a period altogether")).toBe(
-      false,
-    );
-  });
-  it("rejects lines which don't start with uppercase", () => {
-    expect(parser.isTraitStart("this. line Doesn't start with uppercase")).toBe(
-      false,
-    );
-  });
-  it("rejects lines whose final word before the first period isn't uppercase", () => {
-    expect(parser.isTraitStart("This is almost. A trait")).toBe(false);
-  });
-  it("accepts valid starts", () => {
-    expect(parser.isTraitStart("Bound. A secret, mundane")).toBe(true);
-    expect(parser.isTraitStart("Heads. Choose how many heads")).toBe(true);
-  });
-});
 
 describe("parseStatblock", () => {
   it("Does not infinitely loop when it can't find the end of the statblock", () => {
