+++ src/parser.ts
@@ -1,117 +1,6 @@
+import { parseAttacks, type Attack } from './attacks.js';
+import { isTraitStart, parseTraits, type Trait } from './traits.js';
 import { getLines, getName } from './util.js';
-
-export type Attack = {
-  /** The number of this attack that can be made per turn */
-  quantity?: string;
-  /** The name of this attack */
-  name: string;
-  /** The range of this attack */
-  range?: string;
-  /** The to-hit bonus for this attack */
-  bonus?: string;
-  /** The damage this attack does */
-  damage?: string;
-};
-
-/**
- * Parse a single attack
- * @param attack a string describing a single attack
- */
-export function parseAttack(attack: string): Attack {
-  let matches = attack.match(
-    /^(?<qty>\d+) (?<name>.+)( (?<bonus>(\+|-)(\d+)) \((?<damage>.+)\))$/,
-  );
-  if (!matches) {
-    matches = attack.match(/^(?<qty>\d+) (?<name>.*)( (?<bonus>(\+|-)(\d+)))$/);
-  }
-  if (!matches) {
-    matches = attack.match(/^(?<qty>\d+) (?<name>.*)$/);
-  }
-  if (!matches) {
-    matches = attack.match(/^(?<name>.*)$/);
-  }
-
-  const weaponName = matches?.groups?.name?.match(/^[^(]+/)?.[0].trim();
-  const weaponRange = matches?.groups?.name?.match(/\((.+)\)/)?.[1]?.trim();
-
-  if (!weaponName) {
-    throw new Error(`Failed to parse attack: ${attack}`);
-  }
-
-  return {
-    quantity: matches?.groups?.qty,
-    name: weaponName,
-    range: weaponRange,
-    bonus: matches?.groups?.bonus,
-    damage: matches?.groups?.damage,
-  };
-}
-
-/**
- * Parse the attack line of a statblock
- * Group the attacks based on the "and" and "or" usage
- * @param attacks A string of attacks
- * @returns a 2D array of attacks. Each sub-array is a set of attacks
- * that can all happen on the same turn
- */
-export function parseAttacks(attacks: string): Attack[][] {
-  return attacks
-    .split(" or ")
-    .map((group) => group.split(" and ").map((atk) => parseAttack(atk)));
-}
-
-/**
- * Determine whether a line is the beginning of a new trait
- * @param line The line of the statblock
- */
-export function isTraitStart(line: string): boolean {
-  if (line.indexOf(".") === -1) return false;
-  if (!/^[A-Z]$/.test(line.charAt(0))) return false;
-  const lastWordBeforePeriod = line.match(/[a-zA-Z0-9_)-]+\./)?.[0];
-  if (!lastWordBeforePeriod) {
-    throw new Error(`Failed to parse possible trait start, ${line}`);
-  }
-  if (!/^[A-Z]$/.test(lastWordBeforePeriod.charAt(0))) return false;
-  return true;
-}
-
-/**
- * A trait is an aspect of some kind on items, monsters, etc
- */
-export type Trait = {
-  /** The name of the trait */
-  name: string;
-  /** The description of the trait */
-  description: string;
-};
-
-/**
- * Parse the traits section of a statblock
- * @param lines the statblock lines containing traits
- */
-export function parseTraits(lines: string[]): Trait[] {
-  const parsed = [];
-  let current = "";
-  for (let i = 0; i < lines.length; i++) {
-    const line = lines[i]!;
-    if (isTraitStart(line)) {
-      parsed.push(current.trim());
-      current = line;
-    } else {
-      current += ` ${line}`;
-    }
-  }
-  parsed.push(current.trim());
-  return parsed.filter(Boolean).map((trait) => {
-    const indexOfPeriod = trait.indexOf(". ");
-    const name = trait.slice(0, indexOfPeriod).trim();
-    const description = trait.slice(indexOfPeriod + 1, trait.length).trim();
-    return {
-      name,
-      description,
-    };
-  });
-}
 
 type Monster = {
   /** Name */
